compute points on the singular curve over F_8
note that the curve is singular when x=1 since 2y=0 and \del_x g(x,y) = (x+1)^2 over F_2. one should search including the exponent infinity for a^k where a is the generator of F_8 to include 0. one must exclude cases when x=1, i.e. when k=0 or k=7. this yields eight points, though it should be q+1=9 or q-1=7, so not sure what's going on there. searching for a map from the points on the curve over F_2 to the points on the curve over F_7, but still seems illusory.